Route agent-lane fallback through agents context
